+++ src/frontend/src/features/rooms/livekit/components/controls/InfoToggle.tsx
@@ -4,6 +4,7 @@ import { css } from '@/styled-system/css'
 import { ToggleButton } from '@/primitives'
 import { useSidePanel } from '../../hooks/useSidePanel'
 import { ToggleButtonProps } from '@/primitives/ToggleButton'
+import { useSidePanelTriggerRef } from '../../hooks/useSidePanelTriggerRef'
 
 export const InfoToggle = ({
   onPress,
@@ -13,6 +14,7 @@ export const InfoToggle = ({
 
   const { isInfoOpen, toggleInfo } = useSidePanel()
   const tooltipLabel = isInfoOpen ? 'open' : 'closed'
+  const setInfoTriggerRef = useSidePanelTriggerRef('info')
 
   return (
     <div
@@ -27,6 +29,7 @@ export const InfoToggle = ({
         aria-label={t(tooltipLabel)}
         tooltip={t(tooltipLabel)}
         isSelected={isInfoOpen}
+        ref={setInfoTriggerRef}
         onPress={(e) => {
           toggleInfo()
           onPress?.(e)